refactor: move minePlot to its own file
