+++ src/game-logic/reducers/index.test.js
@@ -34,8 +34,6 @@ import {
 import { huggingMachine, sampleCropItem1 } from '../../data/items'
 import { sampleRecipe1 } from '../../data/recipes'
 import { itemsMap } from '../../data/maps'
-import { goldOre } from '../../data/ores'
-import { ResourceFactory } from '../../factories'
 import { genders, standardCowColors, toolType, toolLevel } from '../../enums'
 import {
   farmProductSalesVolumeNeededForLevel,
@@ -2537,71 +2535,5 @@ describe('prependPendingPeerMessage', () => {
     })
 
     expect(pendingPeerMessages).toHaveLength(MAX_PENDING_PEER_MESSAGES)
-  })
-})
-
-describe('minePlot', () => {
-  let gameState
-
-  beforeAll(() => {
-    gameState = {
-      field: [[null, 'crop']],
-      inventory: [],
-      inventoryLimit: 99,
-      toolLevels: {
-        [toolType.SHOVEL]: toolLevel.DEFAULT,
-      },
-    }
-
-    jest.spyOn(ResourceFactory, 'instance')
-
-    ResourceFactory.instance.mockReturnValue({
-      generateResources: () => [goldOre],
-    })
-
-    gameState = fn.minePlot(gameState, 0, 0)
-  })
-
-  test('updates the plot to be shoveled if the plot is empty', () => {
-    expect(gameState.field[0][0].isShoveled).toEqual(true)
-  })
-
-  test('sets the oreId on the plot if ore was spawned', () => {
-    expect(gameState.field[0][0].oreId).toEqual(goldOre.id)
-  })
-
-  test('sets the days until clear', () => {
-    expect(gameState.field[0][0].daysUntilClear > 0).toEqual(true)
-  })
-
-  test('adds the spawned ore to the inventory', () => {
-    let itemIsInInventory = false
-
-    for (let item of gameState.inventory) {
-      if (item.id === goldOre.id) {
-        itemIsInInventory = true
-        break
-      }
-    }
-
-    expect(itemIsInInventory).toEqual(true)
-  })
-
-  test('does not alter the plot if something is already there', () => {
-    expect(gameState.field[0][1]).toEqual('crop')
-  })
-
-  test('shows a notification if there is no room in the inventory', () => {
-    gameState.inventoryLimit = 0
-    gameState.showNotifications = true
-    gameState.todaysNotifications = []
-    gameState.field[0][0] = null
-
-    const { latestNotification } = fn.minePlot(gameState, 0, 0)
-
-    expect(latestNotification).toEqual({
-      message: 'Your inventory is full!',
-      severity: 'info',
-    })
   })
 })
